Share overlay parts between dialog + drawer
